Move location search to components directory for consistency.
